Relay Obfuscator implementation changes
